Issue #435 added new parameters --virtualbox-skip-update --vmwarevsphere-skip-update --vmwarefusion-skip-update
